+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Change Log
 
 ## [Unreleased]
+## [1.1.5] 2026-08-01
 
 ### Fixed